--- addons/survey/static/src/scss/section_backend.scss
@@ -1,13 +1,0 @@
-table.o_section_list_view tr.o_data_row.o_is_line_section {
-    font-weight: bold;
-    background-color: #DDDDDD;
-    border-top: 1px solid #BBB;
-    border-bottom: 1px solid #BBB;
-}
-
-.o_modal_section_question {
-    background-color: #FFFFFF;
-    height: auto;
-    width: auto;
-    padding: 15px;
-}
