+++ addons/survey/views/assets.xml
@@ -25,8 +25,7 @@
     <template id="assets_backend" name="survey assets" inherit_id="web.assets_backend">
         <xpath expr="." position="inside">
             <link rel="stylesheet" href="/survey/static/src/css/survey_result.css"/>
-            <link rel="stylesheet" type="text/scss" href="/survey/static/src/scss/section_backend.scss"/>
-            <script type="text/javascript" src="/survey/static/src/js/section_fields_backend.js"/>
+            <script type="text/javascript" src="/survey/static/src/js/fields_section_one2many.js"/>
         </xpath>
     </template>
