+++ addons/survey/views/survey_question_views.xml
@@ -11,7 +11,6 @@
                 <field name="survey_id" invisible="1"/>
                 <field name="display_type" invisible="1"/>
                 <field name="sequence" invisible="1"/>
-                <div class="o_modal_section_question">
                 <sheet>
                     <div class="oe_title" style="width: 100%;">
                         <label for="title" string="Section" attrs="{'invisible': [('is_page', '=', False)]}"/>
@@ -160,7 +159,6 @@
                         </page>
                     </notebook>
                 </sheet>
-                </div>
             </form>
         </field>
     </record>
