+++ addons/survey/views/survey_survey_views.xml
@@ -54,7 +54,7 @@
                     </group>
                     <notebook>
                         <page string="Questions">
-                            <field name="question_and_page_ids" nolabel="1" widget="section_one2many" mode="tree" context="{'default_survey_id': active_id, 'default_questions_selection': questions_selection}" class="o_survey_questions_and_pages">
+                            <field name="question_and_page_ids" nolabel="1" widget="survey_section_one2many" mode="tree" context="{'default_survey_id': active_id, 'default_questions_selection': questions_selection}" class="o_survey_questions_and_pages">
                                 <tree decoration-bf="is_page==True" editable="bottom">
                                     <control>
                                         <create name="add_section_control" string="Add a section" context="{'default_is_page': True, 'default_questions_selection': 'all', 'default_display_type': 'line_section'}"/>
